requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
colorama==0.4.6
django-hCaptcha==0.2.0
Django==4.0
et-xmlfile==1.1.0
exceptiongroup==1.1.3
idna==3.4
iniconfig==2.0.0
openpyxl==3.1.2
packaging==23.1
Pillow==10.0.0
pip==23.2.1
pluggy==1.2.0
pytest-django==4.5.2
python-dotenv==1.0.0
requests==2.31.0
setuptools==63.2.0
sqlparse==0.4.4
tomli==2.0.1
typing-extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
